+++ app/javascript/vue/tasks/data_attributes/duplicate_data_attributes/main.js
@@ -1,13 +1,14 @@
 import { createApp } from 'vue'
 import App from './app.vue'
 
-function init() {
+function initApp (element) {
   const app = createApp(App)
-  app.mount('#single_bibtex_source')
+
+  app.mount(element)
 }
 
 document.addEventListener('turbolinks:load', () => {
-  if (document.querySelector('#single_bibtex_source')) {
-    init()
-  }
+  const el = document.querySelector('#duplicate_data_attributes_task')
+
+  if (el) { initApp(el) }
 })