+++ CHANGELOG.md
@@ -15,12 +15,14 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 ### Fixed
 
 - Save notice is missing in New type specimen task
+- Fix: Rails now correctly applies time zones for `user_date_start` and `user_date_end` parameters from the `X-Timezone` header [#4721]
 
 ### Changed
 
 - Removed unused BibTeX task [#1440]
 
 [#1440]: https://github.com/SpeciesFileGroup/taxonworks/issues/1440
+[#4721]: https://github.com/SpeciesFileGroup/taxonworks/issues/4721
 [#4739]: https://github.com/SpeciesFileGroup/taxonworks/issues/4739
 
 ## [0.59.0] - 2026-02-04