+++ CHANGELOG.md
@@ -9,12 +9,22 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ### Added
 
+- Report per-object duplicate predicate use [#4741]
+- Batch import BibTeX sources option to include Identifier
 - Batch Upload Asserted Distributions now supports a `gazetteer_id` column [#4739]
 
 ### Fixed
 
 - Save notice is missing in New type specimen task
+- Fix: Rails now correctly applies time zones for `user_date_start` and `user_date_end` parameters from the `X-Timezone` header [#4721]
 
+### Changed
+
+- Removed unused BibTeX task [#1440]
+
+[#4741]: https://github.com/SpeciesFileGroup/taxonworks/issues/4741
+[#1440]: https://github.com/SpeciesFileGroup/taxonworks/issues/1440
+[#4721]: https://github.com/SpeciesFileGroup/taxonworks/issues/4721
 [#4739]: https://github.com/SpeciesFileGroup/taxonworks/issues/4739
 
 ## [0.59.0] - 2026-02-04
@@ -6070,7 +6080,8 @@ _Special thanks to Tom Klein for his amazing open-source contributions on this r
 - Loosing input page numbers when switching tabs on New Taxon Name task
 
 [#1532]: https://github.com/SpeciesFileGroup/taxonworks/issues/1532
-[unreleased]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.58.0...development
+[unreleased]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.59.0...development
+[0.58.0]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.58.0...v0.59.0
 [0.58.0]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.57.0...v0.58.0
 [0.57.0]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.56.1...v0.57.0
 [0.56.1]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.56.0...v0.56.1